+++ lib/Handlers/include/GetQueueCountHandler.hpp
@@ -4,8 +4,12 @@
 
 class GetQueueCountHandler : public Handler, public std::enable_shared_from_this<GetQueueCountHandler> {
 public:
-    bool canHandle(const Request& request) const override;
-    Response handle(const Request& request) const override;
-    std::shared_ptr<Handler> findHandler(const Request& request) override;
+    void handle_request(beast::http::request<beast::http::string_body>& request,
+                        beast::http::response<beast::http::string_body>& response) override {
+        response.result(beast::http::status::ok);
+        response.set(beast::http::field::server, "Beast");
+        response.set(beast::http::field::content_type, "text/html");
+        response.body() = "17";
+        response.prepare_payload();
+    }
 };
-
