+++ lib/Server/include/server.hpp
@@ -1,19 +1,28 @@
 #pragma once
-
-#include <memory>
 #include <boost/asio.hpp>
-#include "handler.hpp"
+#include <memory>
+#include <string>
+
 
 class Server {
 public:
-    Server(int port);
-    void run(std::shared_ptr<Handler> handler);
+    Server(const std::string& address = "127.0.0.1", 
+            unsigned short port = 8080, 
+            const std::string& doc_root = ".",
+            const int threads = 1)
+        : address_(boost::asio::ip::make_address(address)),
+          port_(port),
+          doc_root_(std::make_shared<std::string>(doc_root)),
+          threads_(threads) {}
+
+    void run();
 
 private:
-    boost::asio::io_service io_service_;
-    boost::asio::ip::tcp::acceptor acceptor_;
-    boost::asio::ip::tcp::socket socket_;
-
-    void startAccept();
-    void handleAccept(std::shared_ptr<Handler> handler, const boost::system::error_code& error);
+    // Router& router_;
+    const boost::asio::ip::address address_;
+    const unsigned short port_;
+    const std::shared_ptr<std::string> doc_root_;
+    const int threads_ = 1;
+    // tcp::acceptor acceptor_;
+    // tcp::socket socket_;
 };
