+++ lib/Server/include/session.hpp
@@ -1,20 +1,145 @@
+// Session.hpp
 #pragma once
-
-#include <boost/asio.hpp>
-#include <memory>
 #include "handler.hpp"
+#include "error.hpp"
+#include <boost/beast.hpp>
+#include <boost/asio.hpp>
+#include <iostream>
+#include <memory>
 
-class Session : public std::enable_shared_from_this<Session> {
+// Handles an HTTP server connection
+class session : public std::enable_shared_from_this<session>
+{
+    // This is the C++11 equivalent of a generic lambda.
+    // The function object is used to send an HTTP message.
+    struct send_lambda
+    {
+        session& self_;
+
+        explicit
+        send_lambda(session& self)
+            : self_(self)
+        {
+        }
+
+        template<bool isRequest, class Body, class Fields>
+        void
+        operator()(boost::beast::http::message<isRequest, Body, Fields>&& msg) const
+        {
+            // The lifetime of the message has to extend
+            // for the duration of the async operation so
+            // we use a shared_ptr to manage it.
+            auto sp = std::make_shared<
+                boost::beast::http::message<isRequest, Body, Fields>>(std::move(msg));
+
+            // Store a type-erased version of the shared
+            // pointer in the class to keep it alive.
+            self_.res_ = sp;
+
+            // Write the response
+            boost::beast::http::async_write(
+                self_.stream_,
+                *sp,
+                boost::beast::bind_front_handler(
+                    &session::on_write,
+                    self_.shared_from_this(),
+                    sp->need_eof()));
+        }
+    };
+
+    boost::beast::tcp_stream stream_;
+    boost::beast::flat_buffer buffer_;
+    std::shared_ptr<std::string const> doc_root_;
+    boost::beast::http::request<boost::beast::http::string_body> req_;
+    std::shared_ptr<void> res_;
+    send_lambda lambda_;
+
 public:
-    Session(boost::asio::ip::tcp::socket socket, std::shared_ptr<Handler> handler);
-    void start();
+    // Take ownership of the stream
+    session(
+        boost::asio::ip::tcp::socket&& socket,
+        std::shared_ptr<std::string const> const& doc_root)
+        : stream_(std::move(socket))
+        , doc_root_(doc_root)
+        , lambda_(*this)
+    {
+    }
 
-private:
-    boost::asio::ip::tcp::socket socket_;
-    std::shared_ptr<Handler> handler_;
-    enum { max_length = 1024 };
-    char data_[max_length];
+    // Start the asynchronous operation
+    void
+    run()
+    {
+        do_read();
+    }
 
-    void handleRead(const boost::system::error_code& error, size_t bytes_transferred);
-    void handleWrite(const boost::system::error_code& error);
+    void
+    do_read()
+    {
+        // Make the request empty before reading,
+        // otherwise the operation behavior is undefined.
+        req_ = {};
+
+        // Set the timeout.
+        stream_.expires_after(std::chrono::seconds(30));
+
+        // Read a request
+        boost::beast::http::async_read(stream_, buffer_, req_,
+            boost::beast::bind_front_handler(
+                &session::on_read,
+                shared_from_this()));
+    }
+
+    void
+    on_read(
+        boost::beast::error_code ec,
+        std::size_t bytes_transferred)
+    {
+        boost::ignore_unused(bytes_transferred);
+
+        // This means they closed the connection
+        if(ec == boost::beast::http::error::end_of_stream) {
+            std::cout << "ec == http::error::end_of_stream REQ TARGET" << req_.target() << std::endl;
+            return do_close();
+        }
+        if(ec)
+            return fail(ec, "read");
+
+        // Send the response
+        handle_request(*doc_root_, std::move(req_), lambda_);
+    }
+
+    void
+    on_write(
+        bool close,
+        boost::beast::error_code ec,
+        std::size_t bytes_transferred)
+    {
+        boost::ignore_unused(bytes_transferred);
+
+        if(ec)
+            return fail(ec, "write");
+
+        if(close)
+        {
+            // This means we should close the connection, usually because
+            // the response indicated the "Connection: close" semantic.
+            return do_close();
+        }
+
+        // We're done with the response so delete it
+        res_ = nullptr;
+
+        // Read another request
+        do_read();
+    }
+
+    void
+    do_close()
+    {
+        // Send a TCP shutdown
+        boost::beast::error_code ec;
+        stream_.socket().shutdown(boost::asio::ip::tcp::socket::shutdown_send, ec);
+
+        // At this point the connection is closed gracefully
+    }
 };
