+++ lib/Server/src/server.cpp
@@ -1,32 +1,28 @@
-#include "handler.hpp"
 #include "server.hpp"
-#include "session.hpp"
+#include "listener.hpp"
+#include <boost/asio.hpp>
 
 
-Server::Server(int port) : acceptor_(io_service_, boost::asio::ip::tcp::endpoint(boost::asio::ip::tcp::v4(), port)),
-                           socket_(io_service_) {
-}
+void Server::run() {
+    boost::asio::io_context ioc_{threads_};
 
-void Server::run(std::shared_ptr<Handler> handler) {
-    startAccept();
-    io_service_.run();
-}
+    std::make_shared<listener>(
+        ioc_,
+        boost::asio::ip::tcp::endpoint{address_, port_},
+        doc_root_)->run();
 
-void Server::startAccept() {
-    acceptor_.async_accept(socket_, [this](const boost::system::error_code& error) {
-        if (!error) {
-            auto handler = std::make_shared<Handler>();
-            handleAccept(handler, error);
-        } else {
-            startAccept();
-        }
-    });
-}
 
-void Server::handleAccept(std::shared_ptr<Handler> handler, const boost::system::error_code& error) {
-    if (!error) {
-        std::make_shared<Session>(std::move(socket_), handler)->start();
-    }
+    std::vector<std::thread> v;
+    v.reserve(threads_ - 1);
+    for(auto i = threads_ - 1; i > 0; --i)
+        v.emplace_back(
+        [&ioc_]
+        {
+            ioc_.run();
+        });
+    
 
-    startAccept();
-}
+    std::cout << "Server listening on http://" << address_.to_string() << ":" << port_ << std::endl;
+    ioc_.run();
+    std::cout << "Server stopped" << std::endl;
+};
